Rename State to StateModel
