+++ src/Application/ApplicationMappingProfile.cs
@@ -1,4 +1,5 @@
-﻿using EnduranceJudge.Core.Mappings;
+﻿using EnduranceJudge.Application.State;
+using EnduranceJudge.Core.Mappings;
 using EnduranceJudge.Domain.State;
 using System.Reflection;
 
@@ -8,7 +9,7 @@ public class ApplicationMappingProfile : MappingProfile
 {
     public ApplicationMappingProfile()
     {
-        this.CreateMap<IState, State>();
+        this.CreateMap<IState, StateModel>();
     }
     
     protected override Assembly[] Assemblies => ApplicationConstants.Assemblies;
