+++ src/Application/ApplicationServices.cs
@@ -1,5 +1,6 @@
 ﻿using EnduranceJudge.Application.Core;
 using EnduranceJudge.Application.Services;
+using EnduranceJudge.Application.State;
 using EnduranceJudge.Domain.State;
 using Microsoft.Extensions.DependencyInjection;
 using System.Collections.Generic;
@@ -23,7 +24,7 @@ public static class ApplicationServices
         // services.AddHttpClient();
         // var kur = services.FirstOrDefault(x => x.ServiceType == typeof(IServiceScopeFactory));
 
-        services.AddSingleton<IStateSetter, State>();
+        services.AddSingleton<IStateSetter, StateModel>();
         services.AddTransient<IState>(x => x.GetRequiredService<IStateSetter>());
         return services;
     }
