+++ src/Application/Services/IStateSetter.cs
@@ -1,8 +1,9 @@
-﻿using EnduranceJudge.Domain.State;
+﻿using EnduranceJudge.Application.State;
+using EnduranceJudge.Domain.State;
 
 namespace EnduranceJudge.Application.Services;
 
 public interface IStateSetter : IState
 {
-    internal void Set(State initial);
+    internal void Set(StateModel initial);
 }
