+++ src/Application/State/StateModel.cs
@@ -8,9 +8,9 @@ using EnduranceJudge.Domain.State.Participations;
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace EnduranceJudge.Application;
+namespace EnduranceJudge.Application.State;
 
-public class State : IStateSetter
+public class StateModel : IStateSetter
 {
     public EnduranceEvent Event { get; set; }
     public List<Horse> Horses { get; private set; } = new();
@@ -20,7 +20,7 @@ public class State : IStateSetter
     [JsonIgnore]
     public IReadOnlyList<Country> Countries
         => ApplicationConstants.Countries.List.AsReadOnly();
-    void IStateSetter.Set(State initial)
+    void IStateSetter.Set(StateModel initial)
     {
         this.Event = initial.Event;
         this.Horses = initial.Horses;
